+++ src/api.rs
@@ -547,7 +547,7 @@ impl Engine {
         scripts: &[&str],
         optimization_level: OptimizationLevel,
     ) -> Result<AST, ParseError> {
-        let stream = lex(scripts);
+        let stream = lex(scripts, self.max_string_size);
         self.parse(&mut stream.peekable(), scope, optimization_level)
     }
 
@@ -669,7 +669,7 @@ impl Engine {
 
         // Trims the JSON string and add a '#' in front
         let scripts = ["#", json.trim()];
-        let stream = lex(&scripts);
+        let stream = lex(&scripts, self.max_string_size);
         let ast =
             self.parse_global_expr(&mut stream.peekable(), &scope, OptimizationLevel::None)?;
 
@@ -750,7 +750,7 @@ impl Engine {
         script: &str,
     ) -> Result<AST, ParseError> {
         let scripts = [script];
-        let stream = lex(&scripts);
+        let stream = lex(&scripts, self.max_string_size);
 
         {
             let mut peekable = stream.peekable();
@@ -904,7 +904,7 @@ impl Engine {
         script: &str,
     ) -> Result<T, Box<EvalAltResult>> {
         let scripts = [script];
-        let stream = lex(&scripts);
+        let stream = lex(&scripts, self.max_string_size);
 
         let ast = self.parse_global_expr(
             &mut stream.peekable(),
@@ -1034,7 +1034,7 @@ impl Engine {
         script: &str,
     ) -> Result<(), Box<EvalAltResult>> {
         let scripts = [script];
-        let stream = lex(&scripts);
+        let stream = lex(&scripts, self.max_string_size);
 
         let ast = self.parse(&mut stream.peekable(), scope, self.optimization_level)?;
         self.consume_ast_with_scope(scope, &ast)